+.道具使用实现及玩家信息UI添加
+.道具使用的实现,并添加2个道具功能实现及4个道具
+.实现修改金币数量及在一定范围内随机修改数量的道具(常用语添加金币和添加随机金币)
+.添加小中大金币包和一个随机金币包
+.添加玩家信息UI展示,展示名称和金币数量
#.优化消耗品类型,分为游戏,战斗,通用三种.
#.细节优化及其他实现

+++ Source/GUAO_TBS/Public/GameProps/GameCapabilities/MoneyOffsetCapabilities.h
@@ -19,6 +19,6 @@ public:
 	virtual void OnUseGameCapabilities() override;
 	
 protected:
-	float MoneyOffset;
+	int32 MoneyOffset;
 	
 };
